+++ src/test/java/com/jelly/zzirit/domain/item/TimeDealControllerTest.java
@@ -16,8 +16,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jelly.zzirit.domain.item.dto.timeDeal.request.TimeDealCreateRequest;
-import com.jelly.zzirit.domain.item.dto.timeDeal.response.TimeDealModalCreateResponse;
+import com.jelly.zzirit.domain.item.dto.request.TimeDealCreateRequest;
+import com.jelly.zzirit.domain.item.dto.response.timeDeal.TimeDealModalCreateResponse;
 import com.jelly.zzirit.domain.member.entity.authenum.Role;
 import com.jelly.zzirit.global.security.util.JwtUtil;
 import com.jelly.zzirit.testutil.TimeDealTestHelper;
